# internal/generated/server.go
// Package generated provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package generated

import (
	"errors"
	"fmt"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
)

// Defines values for ValidationErrorDetailCode.
const (
	INVALIDFORMAT ValidationErrorDetailCode = "INVALID_FORMAT"
	REQUIRED      ValidationErrorDetailCode = "REQUIRED"
	TOOLARGE      ValidationErrorDetailCode = "TOO_LARGE"
	TOOLONG       ValidationErrorDetailCode = "TOO_LONG"
	TOOSHORT      ValidationErrorDetailCode = "TOO_SHORT"
	TOOSMALL      ValidationErrorDetailCode = "TOO_SMALL"
)

// Valid indicates whether the value is a known member of the ValidationErrorDetailCode enum.
func (e ValidationErrorDetailCode) Valid() bool {
	switch e {
	case INVALIDFORMAT:
		return true
	case REQUIRED:
		return true
	case TOOLARGE:
		return true
	case TOOLONG:
		return true
	case TOOSHORT:
		return true
	case TOOSMALL:
		return true
	default:
		return false
	}
}

// CreatePostRequest defines model for CreatePostRequest.
type CreatePostRequest struct {
	Category string    `json:"category"`
	Content  string    `json:"content"`
	Tags     *[]string `json:"tags,omitempty"`
	Title    string    `json:"title"`
}

// HealthResponse defines model for HealthResponse.
type HealthResponse struct {
	Status string `json:"status"`
}

// Post defines model for Post.
type Post struct {
	Category  string    `json:"category"`
	Content   string    `json:"content"`
	CreatedAt time.Time `json:"createdAt"`
	Id        int       `json:"id"`
	Tags      *[]string `json:"tags,omitempty"`
	Title     string    `json:"title"`
	UpdatedAt time.Time `json:"updatedAt"`
}

// ProblemDetails defines model for ProblemDetails.
type ProblemDetails struct {
	// Detail Human-readable explanation
	Detail string `json:"detail"`

	// Errors Validation errors (only for 422)
	Errors *[]ValidationErrorDetail `json:"errors,omitempty"`

	// Status HTTP status code
	Status int `json:"status"`

	// Title Short human-readable summary
	Title string `json:"title"`

	// Type Problem type URI
	Type string `json:"type"`
}

// UpdatePostRequest defines model for UpdatePostRequest.
type UpdatePostRequest struct {
	Category string    `json:"category"`
	Content  string    `json:"content"`
	Tags     *[]string `json:"tags,omitempty"`
	Title    string    `json:"title"`
}

// ValidationErrorDetail defines model for ValidationErrorDetail.
type ValidationErrorDetail struct {
	// Code Error code
	Code ValidationErrorDetailCode `json:"code"`

	// Field Name of the field that failed validation
	Field string `json:"field"`

	// Message Human-readable error description
	Message string `json:"message"`
}

// ValidationErrorDetailCode Error code
type ValidationErrorDetailCode string

// PostId defines model for PostId.
type PostId = int

// InternalError defines model for InternalError.
type InternalError = ProblemDetails

// NotFound defines model for NotFound.
type NotFound = ProblemDetails

// ValidationError defines model for ValidationError.
type ValidationError = ProblemDetails

// ListPostsParams defines parameters for ListPosts.
type ListPostsParams struct {
	// Term Search filter — matches against title, content, and category
	Term *string `form:"term,omitempty" json:"term,omitempty"`
}

// CreatePostJSONRequestBody defines body for CreatePost for application/json ContentType.
type CreatePostJSONRequestBody = CreatePostRequest

// UpdatePostJSONRequestBody defines body for UpdatePost for application/json ContentType.
type UpdatePostJSONRequestBody = UpdatePostRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// HealthCheck Health check
	// (GET /health)
	HealthCheck(w http.ResponseWriter, r *http.Request)
	// ListPosts List all posts
	// (GET /posts)
	ListPosts(w http.ResponseWriter, r *http.Request, params ListPostsParams)
	// CreatePost Create a blog post
	// (POST /posts)
	CreatePost(w http.ResponseWriter, r *http.Request)
	// DeletePost Delete a blog post
	// (DELETE /posts/{id})
	DeletePost(w http.ResponseWriter, r *http.Request, id PostId)
	// GetPost Get a blog post
	// (GET /posts/{id})
	GetPost(w http.ResponseWriter, r *http.Request, id PostId)
	// UpdatePost Update a blog post
	// (PUT /posts/{id})
	UpdatePost(w http.ResponseWriter, r *http.Request, id PostId)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// HealthCheck Health check
// (GET /health)
func (_ Unimplemented) HealthCheck(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ListPosts List all posts
// (GET /posts)
func (_ Unimplemented) ListPosts(w http.ResponseWriter, r *http.Request, params ListPostsParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// CreatePost Create a blog post
// (POST /posts)
func (_ Unimplemented) CreatePost(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// DeletePost Delete a blog post
// (DELETE /posts/{id})
func (_ Unimplemented) DeletePost(w http.ResponseWriter, r *http.Request, id PostId) {
	w.WriteHeader(http.StatusNotImplemented)
}

// GetPost Get a blog post
// (GET /posts/{id})
func (_ Unimplemented) GetPost(w http.ResponseWriter, r *http.Request, id PostId) {
	w.WriteHeader(http.StatusNotImplemented)
}

// UpdatePost Update a blog post
// (PUT /posts/{id})
func (_ Unimplemented) UpdatePost(w http.ResponseWriter, r *http.Request, id PostId) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// HealthCheck operation middleware
func (siw *ServerInterfaceWrapper) HealthCheck(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.HealthCheck(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListPosts operation middleware
func (siw *ServerInterfaceWrapper) ListPosts(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params ListPostsParams

	// ------------- Optional query parameter "term" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "term", r.URL.Query(), &params.Term, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "term"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "term", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListPosts(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreatePost operation middleware
func (siw *ServerInterfaceWrapper) CreatePost(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreatePost(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// DeletePost operation middleware
func (siw *ServerInterfaceWrapper) DeletePost(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id PostId

	err = runtime.BindStyledParameterWithOptions("simple", "id", chi.URLParam(r, "id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "", ValueIsUnescaped: r.URL.RawPath == ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DeletePost(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetPost operation middleware
func (siw *ServerInterfaceWrapper) GetPost(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id PostId

	err = runtime.BindStyledParameterWithOptions("simple", "id", chi.URLParam(r, "id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "", ValueIsUnescaped: r.URL.RawPath == ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetPost(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// UpdatePost operation middleware
func (siw *ServerInterfaceWrapper) UpdatePost(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id PostId

	err = runtime.BindStyledParameterWithOptions("simple", "id", chi.URLParam(r, "id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "", ValueIsUnescaped: r.URL.RawPath == ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.UpdatePost(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/health", wrapper.HealthCheck)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/posts", wrapper.ListPosts)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/posts", wrapper.CreatePost)
	})
	r.Group(func(r chi.Router) {
		r.Delete(options.BaseURL+"/posts/{id}", wrapper.DeletePost)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/posts/{id}", wrapper.GetPost)
	})
	r.Group(func(r chi.Router) {
		r.Put(options.BaseURL+"/posts/{id}", wrapper.UpdatePost)
	})

	return r
}